+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 1.0.16
 * Removed 1002 from the retriable exceptions list as it's a catch-all for all other exceptions
+* Change Java Client to 0.6.0-patch3 and switch tests to use same Base class
 
 ## 1.0.15
 * Added code 107 to the exception list and added more DLQ logging